Remove UIKit screens
