--- DNAConverter/DNAConverter/Source/en.lproj/Main.strings
@@ -1,30 +1,0 @@
-
-/* Class = "UIBarButtonItem"; title = "閉じる"; ObjectID = "0gj-jA-S9K"; */
-"0gj-jA-S9K.title" = "Close";
-
-/* Class = "UITextView"; text = "Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda."; ObjectID = "MUd-Bp-mSN"; */
-"MUd-Bp-mSN.text" = "Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.";
-
-/* Class = "UIBarButtonItem"; title = "Clear"; ObjectID = "MqP-WI-HMs"; */
-"MqP-WI-HMs.title" = "Clear";
-
-/* Class = "UIBarButtonItem"; title = "変換する"; ObjectID = "O7E-Jy-jCT"; */
-"O7E-Jy-jCT.title" = "Convert";
-
-/* Class = "UITextView"; text = "Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda."; ObjectID = "Vmg-FA-pp0"; */
-"Vmg-FA-pp0.text" = "Lorem ipsum dolor sit er elit lamet, consectetaur cillium adipisicing pecu, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum. Nam liber te conscient to factor tum poen legum odioque civiuda.";
-
-/* Class = "UINavigationItem"; title = "履歴"; ObjectID = "h8N-w2-AO3"; */
-"h8N-w2-AO3.title" = "History";
-
-/* Class = "UISegmentedControl"; hNr-8C-VBC.segmentTitles[0] = "ことば"; ObjectID = "hNr-8C-VBC"; */
-"hNr-8C-VBC.segmentTitles[0]" = "Language";
-
-/* Class = "UISegmentedControl"; hNr-8C-VBC.segmentTitles[1] = "ぬくれおちど"; ObjectID = "hNr-8C-VBC"; */
-"hNr-8C-VBC.segmentTitles[1]" = "Nucleotide";
-
-/* Class = "UINavigationItem"; title = "DNA変換"; ObjectID = "hve-Cv-6Da"; */
-"hve-Cv-6Da.title" = "DNA Converter";
-
-/* Class = "UILabel"; text = "履歴がありません"; ObjectID = "moW-ph-SFy"; */
-"moW-ph-SFy.text" = "No history";
