--- DNAConverter/DNAConverter/Source/ja.lproj/LaunchScreen.strings
@@ -1,1 +1,0 @@
-
